requirements.txt
aiohttp
chromadb
crawl4ai
faiss-cpu
fastapi
feedparser
langchain-openai
# librosa
markdownify
mcp
openai
pillow
pyflakes
numpy
scipy
matplotlib
pydantic
pydub
PyPDF2
pygame
pyrubberband
python-magic
python-multipart
requests
rich
scikit-learn
soundfile
urllib3
uvicorn
yt-dlp
youtube-transcript-api
psola
